requirements.txt
google-generativeai
requests
beautifulsoup4
python-dotenv
flask
flask_cors
gunicorn